api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
authors = [
    {name = "Christopher Bonner"},
{name = "Freddie Crumpton"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "boto3 (>=1.37.31,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "botocore (>=1.38.7,<2.0.0)",
    "rsa (>=4.9.1,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pycryptodome (>=3.22.0,<4.0.0)",
    "flask-socketio (>=5.5.1,<6.0.0)",
    "eventlet (>=0.39.1,<0.40.0)",
    "ffmpeg (>=1.4,<2.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "faker (>=37.3.0,<38.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
